fix(floating-toolbar): exclude event listener from toolbar
fix #4370
