+++ src/views/ArticleDetail/index.tsx
@@ -5,7 +5,6 @@ import { useContext, useEffect, useState } from 'react'
 import { FormattedMessage } from 'react-intl'
 
 import {
-  KEYVALUE,
   OPEN_COMMENT_DETAIL_DIALOG,
   OPEN_COMMENT_LIST_DRAWER,
 } from '~/common/enums'
@@ -33,7 +32,6 @@ import {
   useEventListener,
   useFeatures,
   useIntersectionObserver,
-  useNativeEventListener,
   usePublicQuery,
   useRoute,
   ViewerContext,
@@ -540,30 +538,6 @@ const ArticleDetail = ({
   useEffect(() => {
     loadPrivate()
   }, [article?.shortHash, viewer.id])
-
-  // floating toolbar is blocking some text when pressing page up and down
-  // offsetting it with the height of floating toolbar
-  useNativeEventListener('keydown', (event: KeyboardEvent) => {
-    const keyToScrollDirection = {
-      [KEYVALUE.pageDown]: 1,
-      [KEYVALUE.pageUp]: -1,
-      [KEYVALUE.space]: 1,
-    } // map the key to scroll direction
-    const key = event.code.toLowerCase()
-    if (key in keyToScrollDirection) {
-      event.preventDefault()
-      const remInPixels = parseFloat(
-        getComputedStyle(document.documentElement).fontSize
-      )
-      const scrollDirection = keyToScrollDirection[key]
-      const scrollAmount = window.innerHeight - 5 * remInPixels // the height of floating toolbar
-
-      window.scrollBy({
-        top: scrollAmount * scrollDirection,
-        behavior: 'smooth',
-      })
-    }
-  })
 
   /**
    * Render:Loading
